requirements.txt
pytest==7.4.3
requests==2.31.0
allure-pytest==2.13.2
PyYAML==6.0.1
pytest-html==4.1.1
openpyxl==3.1.2
python-dotenv==1.0.0